requirements.txt
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorlog==6.10.1
gitdb==4.0.12
GitPython==3.1.46
hydra-colorlog==1.2.0
hydra-core==1.3.2
idna==3.11
omegaconf==2.3.0
packaging==26.0
platformdirs==4.5.1
protobuf==6.33.5
pydantic==2.12.5
pydantic_core==2.41.5
PyYAML==6.0.3
requests==2.32.5
sentry-sdk==2.52.0
smmap==5.0.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
wandb==0.24.2
